requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
argcomplete==3.2.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.7
botocore==1.34.7
certifi==2023.11.17
cffi==1.16.0
cfn-flip==1.3.0
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0
django-cors-headers==4.3.1
django-s3-storage==0.15.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
durationpy==0.6
frozenlist==1.4.0
googletrans==3.1.0a0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.9.0
h2==3.2.0
hjson==3.1.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
mysql-client==0.0.1
packaging==23.2
Pillow==10.1.0
placebo==0.9.0
promise==2.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.3.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
rfc3986==1.5.0
s3transfer==0.10.0
service-identity==23.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.3
twilio==8.11.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.26.0
Werkzeug==3.0.1
wheel==0.43.0
yarl==1.9.4
zappa==0.58.0
zope.interface==6.1
